pyproject.toml
[project]
name = "1099-validator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "boto3>=1.36.6",
    "fastapi>=0.115.7",
    "langchain-community>=0.3.15",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
]
